refactor: split dedup image pixel match logic into dedicated file and separate tests
